+++ spec/controllers/groups_controller_spec.rb
@@ -714,9 +714,11 @@ describe GroupsController do
     context 'a student' do
       let(:user) { create :student }
       let(:assignment) { create :assignment }
-      let!(:starter_file_group) { create :starter_file_group_with_entries, assignment: assignment }
-      let!(:grouping) { create :grouping_with_inviter, assignment: assignment, inviter: user }
+      let(:starter_file_group) { create :starter_file_group_with_entries, assignment: assignment }
+      let(:grouping) { create :grouping_with_inviter, assignment: assignment, inviter: user }
       it 'should send a zip file containing the correct content' do
+        starter_file_group
+        grouping
         expect(controller).to receive(:send_file) do |file_path|
           Zip::File.open(Rails.root + file_path) do |zipfile|
             expect(zipfile.entries.map(&:name)).to contain_exactly('q1/', 'q1/q1.txt', 'q2.txt')
@@ -725,6 +727,21 @@ describe GroupsController do
           end
         end
         subject
+      end
+
+      context 'when the grouping was created before any starter file groups' do
+        it 'should send a zip file containing the correct content' do
+          grouping
+          starter_file_group
+          expect(controller).to receive(:send_file) do |file_path|
+            Zip::File.open(Rails.root + file_path) do |zipfile|
+              expect(zipfile.entries.map(&:name)).to contain_exactly('q1/', 'q1/q1.txt', 'q2.txt')
+              expect(zipfile.find_entry('q1/q1.txt').get_input_stream.read.strip).to eq 'q1 content'
+              expect(zipfile.find_entry('q2.txt').get_input_stream.read.strip).to eq 'q2 content'
+            end
+          end
+          subject
+        end
       end
     end
   end