+++ Changelog.md
@@ -17,6 +17,7 @@
 - Support Markdown syntax for assessment messages (#5135)
 - Remove controller-specific css files (#5136)
 - Replace non-UTF8 characters in text file preview (#5156)
+- Rollback group creation if error is raised when creating a group for a working-alone student (#5169)
 - Prevent deletion/modification of annotation texts associated with a result with a pending remark request (#5170)
 
 ## [v1.11.4]